Resolve NeMo Gym review threads
