+++ packages/tasksets/tasksets/utils/nemo_gym_utils.py
@@ -1,18 +1,8 @@
 from pathlib import Path
-from typing import TypeAlias, cast
 
-ConfigMap: TypeAlias = dict[str, object]
+from verifiers.v1.utils.nemo_gym_utils import nemo_gym_package_root
+
 DEFAULT_NEMO_GYM_DATA_NAME = "example.jsonl"
-
-
-def nemo_gym_package_root() -> Path:
-    try:
-        from nemo_gym import PARENT_DIR as nemo_gym_root  # ty: ignore[unresolved-import]
-    except ImportError as exc:
-        raise ImportError(
-            "NeMo Gym integration requires nemo-gym. Install as `verifiers[nemogym]`."
-        ) from exc
-    return Path(nemo_gym_root)
 
 
 def resolve_nemo_gym_data_path(
@@ -23,10 +13,3 @@ def resolve_nemo_gym_data_path(
     if not path.exists():
         raise FileNotFoundError(f"NeMo Gym data file not found: {path}")
     return path
-
-
-def agent_ref_name(value: object) -> str | None:
-    if not isinstance(value, dict):
-        return None
-    name = cast(ConfigMap, value).get("name")
-    return name if isinstance(name, str) and name else None
